Refactor(src): move processing functions to `genai` module
